superpool/pyproject.toml
[tool.poetry]
name = "superpool"
version = "0.1"
description = "Insurance Aggregation Platform for Financial Institutions"
authors = ["unyte-africa <support@unyte.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.1"
djangorestframework = "^3.15.1"
django-cors-headers = "^4.3.1"
djangorestframework-simplejwt = "^5.3.1"
djangorestframework-api-key = "^3.0.0"
django-two-factor-auth = "^1.16.0"
django-environ = "^0.11.2"
drf-spectacular = "^0.27.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
faker = "^25.4.0"
factory-boy = "^3.3.0"
django-extensions = "^3.2.3"
ipdb = "^0.13.13"
watchman = "^0.0.1"
django-debug-toolbar = "^4.4.2"
rich = "^13.7.1"
psycopg = "^3.1.19"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
